+++ packages/widgetbook/lib/src/layout/base_layout.dart
@@ -8,13 +8,11 @@ abstract class BaseLayout {
     required this.navigationBuilder,
     required this.addonsBuilder,
     required this.knobsBuilder,
-    required this.argsBuilder,
     required this.workbench,
   });
 
   final Widget Function(BuildContext context) navigationBuilder;
   final List<Widget> Function(BuildContext context) addonsBuilder;
   final List<Widget> Function(BuildContext context) knobsBuilder;
-  final List<Widget> Function(BuildContext context) argsBuilder;
   final Widget workbench;
 }