+++ packages/widgetbook/lib/src/next/addons/base/addon.dart
@@ -1,6 +1,0 @@
-// next version of Widgetbook doesn't have api docs yet
-// ignore_for_file: public_member_api_docs
-
-import '../../../addons/addons.dart';
-
-typedef Addon<T> = WidgetbookAddon<T>;