+++ packages/widgetbook_generator/CHANGELOG.md
@@ -1,3 +1,7 @@
+## Unreleased
+
+- **BREAKING**: Remove `next` builders. ([#1634](https://github.com/widgetbook/widgetbook/pull/1634))
+
 ## 3.17.0
 
 - **BREAKING**: Require `analyzer` >=8.0.0, `build` >=4.0.0 and `source_gen` >=4.0.0. ([#1584](https://github.com/widgetbook/widgetbook/pull/1584))